+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "fifth_element_projectteam",
+  "version": "1.0.0",
+  "description": "This is the start of our groups 301 Final Project.",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nacerillo/Fifth_Element_ProjectTeam.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/nacerillo/Fifth_Element_ProjectTeam/issues"
+  },
+  "homepage": "https://github.com/nacerillo/Fifth_Element_ProjectTeam#readme",
+  "dependencies": {
+    "dotenv": "^8.2.0",
+    "ejs": "^3.1.6",
+    "express": "^4.17.1",
+    "pg": "^8.5.1",
+    "superagent": "^6.1.0"
+  }
+}
