+++ views/pages/index.ejs
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <form action="/cities" method="POST">
+  
+            <input type="text" name="city_name">
+             <input type="submit">
+    
+      </form>
+</body>
+</html>
